consumer added successfully

+++ PayCoreFinalProject.Base/Offer/OfferResponse.cs
@@ -5,7 +5,7 @@ namespace PayCoreFinalProject.Base.Offer;
 public class OfferResponse
 {
     public int ProductId { get; set; }
-    public  double OfferedPrice { get; set; }
+    public  decimal OfferedPrice { get; set; }
     public  ProductDto Product { get; set; }
