+++ PayCoreFinalProject.Dto/OfferDto.cs
@@ -3,7 +3,7 @@ namespace PayCoreFinalProject.Dto;
 public class OfferDto
 {
     public int Id { get; set; }
-    public  double OfferedPrice { get; set; }
+    public  decimal OfferedPrice { get; set; }
     public  int OfferedById { get; set; }
